Comment out Galil for basic build
Galil requires c++11, and so, doesn't build on current vxWorks cross-compilers. They also don't provide any easy configuration options to turn off compilation for vxWorks, so it will be disabled by default to allow for users to easily compile, but it you are only going to use linux or windows, it's available to uncomment.